+++ mod3rd/admin_esp/wlan.py
@@ -2,25 +2,30 @@
 import ubinascii
 
 from modcore.log import logger
-from moddev.wlan import wlan_ap
+from modcore import modc
+from moddev.wlan import wlan_ap, WLAN_RESTART
 from modext.windup import Router
 
 router = Router( root="/admin" )
 
 # get request
 
+netw = {}
+
 @router.get("/wlan")
 def my_form( req, args ):
     
     networks = wlan_ap.scan()
     
-    netw = []
     netwdiv = ""
     
     for nam, mac, _, _, _, _ in networks:
         
         nam = nam.decode() 
         mac = ubinascii.hexlify(mac).decode()
+        
+        global netw
+        netw[mac]=nam
         
         netwdiv += "<div>"
         netwdiv += '<input type="radio" id="f_'+mac+'" name="fwifi" value="'+mac+'" >'
@@ -32,7 +37,11 @@ def my_form( req, args ):
             <html>
             <body>
 
-            <h2>WLAN config</h2>
+            <h2>WLAN configuration</h2>
+
+            <div> &nbsp; </div>
+            <div> Currently connected to: '%s' </div>
+            <div> &nbsp; </div>
 
             <form action="/admin/wlan" method="POST">
             
@@ -45,11 +54,14 @@ def my_form( req, args ):
                 <input type="text" id="f_passwd" name="fpasswd" value=""><br>
 
                 <input type="submit" value="Connect">
+                <div> Make sure to connect via SoftAP for setting up WLAN
+                        otherwise Connection might get lost.
+                </div>
             </form> 
 
             </body>
             </html>            
-            """ % ( netwdiv )
+            """ % ( str(wlan_ap.ssid), netwdiv )
     
     logger.info(data)
     req.send_response( response=data )
@@ -60,14 +72,21 @@ def my_form( req, args ):
 @router.post("/wlan")
 def my_form( req, args ):
     
-    body = req.request.xform
+    form = req.request.xform
+    
+    ssid = netw[ form[ "fwifi" ]]
+    passwd = form[ "fpasswd" ]
     
     data = """
-            <h1>WLAN config data</h1>
-            <div> query parameter = %s </div>
-            <div> post data = %s </div>
-            <div> post type = %s </div>
-            """ % ( repr( args ), repr( body ), type( body ) )
+            <h1>WLAN configuration saved</h1>
+            <div> &nbsp; </div>
+            <div> Networkname: '%s' </div>
+            """ % ( ssid )
+    
+    wlan_ap.wlan_config( ssid, passwd )
+    
+    # restart in next loop, otherwise connection breaks and response fails
+    modc.fire_event(WLAN_RESTART)
     
     logger.info(data)
     req.send_response( response=data )